chore(app): reorganize structure
