+++ src/renderer/login/index.js
@@ -1,0 +1,1 @@
+export { default as Login } from './components'; // eslint-disable-line import/prefer-default-export
