+++ src/renderer/register/actions/registerCompletionActions.js
