+++ src/renderer/register/actions/registerFormActions.js
