+++ src/renderer/register/components/AccountView/LedgerView/LedgerView.scss
