+++ src/renderer/register/components/AccountView/MnemonicView/MnemonicView.scss
