+++ src/renderer/register/components/AccountView/MnemonicView/MnemonicWord/MnemonicWord.scss
