+++ src/renderer/register/components/AccountView/MnemonicView/MnemonicWord/index.js
