+++ src/renderer/register/components/AccountView/MnemonicView/index.js
