+++ src/renderer/register/components/CreateAccount/RegisterForm/RegisterForm.scss
