+++ src/renderer/register/components/CreateAccount/index.js
