+++ src/renderer/register/components/Register/Register.js
