+++ src/renderer/register/components/Register/Register.scss
