+++ src/renderer/register/components/Register/index.js
