+++ src/renderer/register/components/VerifyAccount/VerifyAccount.scss
