+++ src/renderer/register/components/index.js
