+++ src/renderer/register/index.js
@@ -1,0 +1,1 @@
+export { default as Register } from './components'; // eslint-disable-line import/prefer-default-export
