+++ src/renderer/register/shapes/registerShape.js
