+++ src/renderer/shared/components/NavigationButtons/NavigationButtons.scss
