+++ src/renderer/shared/components/SidePanel/SidePanel.scss
