+++ src/renderer/shared/components/SidePanel/index.js
